Fix remaining compilation errors in Phase 20+ features
This commit resolves all high and medium priority compilation errors
documented in COMPILATION_ISSUES.md:

1. Ship Model Field Name Fixes (capture/manager.go):
   - Replace CurrentShields/CurrentHull with Shields/Hull
   - Replace PlayerID with OwnerID
   - Add ShipType lookups for max values and CargoSpace

2. API Converter Type Fixes (converters.go):
   - Fix Speed field: int32 → float64
   - Fix Weapon converter: use WeaponType, RangeValue, proper types
   - Fix Outfit converter: use OutfitType and Modifiers

3. Type Conversion Fixes (shipsystems/manager.go):
   - Add float64() conversions for int→float64 comparisons
   - Add int() conversions for float64→int assignments

4. Cleanup Unused Imports:
   - Remove unused "math/rand" from arena/manager.go
   - Remove unused "models" from marketplace/manager.go
   - Remove unused "models" from mining/manager.go

5. Cleanup Unused Variables:
   - Comment out unused researchCost in manufacturing/manager.go

All affected packages now compile successfully. Remaining TUI errors
are separate issues not covered in the original cleanup scope.

+++ internal/shipsystems/manager.go
@@ -226,12 +226,12 @@ func (m *Manager) ActivateCloak(ctx context.Context, shipID uuid.UUID, ship *mod
 	}
 
 	// Check energy (using shields as energy for now)
-	if ship.Shields < m.config.CloakActivationCost {
+	if float64(ship.Shields) < m.config.CloakActivationCost {
 		return fmt.Errorf("insufficient energy (need %.1f)", m.config.CloakActivationCost)
 	}
 
 	// Deduct activation cost
-	ship.Shields -= m.config.CloakActivationCost
+	ship.Shields -= int(m.config.CloakActivationCost)
 	if err := m.shipRepo.Update(ctx, ship); err != nil {
 		return fmt.Errorf("failed to update ship: %v", err)
 	}
@@ -422,12 +422,12 @@ func (m *Manager) ExecuteJump(ctx context.Context, shipID, fromSystemID, toSyste
 
 	// Calculate fuel cost
 	fuelCost := distance * m.config.JumpDriveFuelCost
-	if ship.Fuel < fuelCost {
+	if float64(ship.Fuel) < fuelCost {
 		return fmt.Errorf("insufficient fuel (need %.1f)", fuelCost)
 	}
 
 	// Deduct fuel
-	ship.Fuel -= fuelCost
+	ship.Fuel -= int(fuelCost)
 	if err := m.shipRepo.Update(ctx, ship); err != nil {
 		return fmt.Errorf("failed to update ship: %v", err)
 	}
